% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_length.R
\name{AgeAssign}
\alias{AgeAssign}
\title{Assign Age from Sampled Age-Length Key.}
\usage{
AgeAssign(len, size, alk, setSeed = NULL)
}
\arguments{
\item{len}{Character vector of lengths.}

\item{size}{Integer vector of length frequencies.}

\item{alk}{Table or List of age-length frequencies.}

\item{setSeed}{Numeric. Default NULL for `sample`.}
}
\value{
Numeric vector of ages.
}
\description{
Using length data & known age-length key, assign age
   based on length data. Can use an age-length key covering range
   of ages and lengths to provide ages to all lengths.
}
\examples{
# coming soon.
}
