% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap-methods.R
\name{ApplyStuRecap}
\alias{ApplyStuRecap}
\title{Conveniently Apply \code{StuRecap} Methods.}
\usage{
ApplyStuRecap(x, ...)
}
\arguments{
\item{x}{Object of class `list` with each element of class `RecapData`
(e.g., \code{TagRecaptures[["Data"]]}).}

\item{...}{Passed on to other methods. See details.}
}
\value{
List with the following items;
   \describe{
      \item{GetRecapInfo}{function to get output of \code{StuRecap.RecapData}}
      \item{GetMatrixData}{function to get 2-column matrix - column 1
            is release year & column 2 is recapture year}
      \item{CrossTab}{cross tab table made from matrix data
            release x recapture}
   }
}
\description{
\code{ApplyStuRecap()} is convenient wrapper for
   \code{StuRecap} methods. It works with the list outputs of such
   methods so user does not have to run for loops or nested
   \code{lapply} functions. See \code{\link{StuRecap}}.
}
\details{
\code{StuRecap.StuRecap()} accepts three arguments: `species`;
   `lencat`; & `minDAL`. Each are described below.
   \describe{
      \item{species}{default to 'w' (White Sturgeon)}
      \item{lencat}{default to `NULL`; choices: s; l; o. defines the
            length category at tagging s = sub-legal, l = legal, and
            o = over-legal}
      \item{minDAL}{minimum days-at-large, default -1 to include all but
            useful for limiting fish caught within a day or week of
            being tagged, if so desired}
   }

   Best to keep \code{species} as default as 'g' (Green Sturgeon;
   the other choice have but ~3 recaptures in entire dataset)
}
\examples{
# coming soon.
}
