% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_ratio.R
\name{ApportionSex}
\alias{ApportionSex}
\title{Apportion Sex (as m or f) for Striped Bass Dataset.}
\usage{
ApportionSex(data, sex)
}
\arguments{
\item{data}{Dataframe containing sex data.}

\item{sex}{Character (quoted or not). The field name in `data`.}
}
\value{
`data` with un-sexed fish now either as 'm' or 'f'.
}
\description{
Striped Bass data analyses are bifurcated by sex. Not all
   fish are or can be sexed in field. So, this function apportions sex
   based on proportion of sexed fish. No fractions, this function assigns
   the 'leftover' 1 to females.
}
\examples{
# coming soon.
}
