% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{Axis.plot_empty}
\alias{Axis.plot_empty}
\title{Display Custom Axis for \code{Plot}.}
\usage{
\method{Axis}{plot_empty}(
  x,
  ...,
  side,
  cexAxis = 1,
  labelAdj = 1,
  interval = NULL,
  offset = 0,
  format = FALSE
)
}
\arguments{
\item{x}{An object of class `plot_empty`.}

\item{...}{Curently not used.}

\item{side}{Integer either 1, 2, 3, or 4 (bottom, left, top, right).}

\item{cexAxis}{Numeric denoting size of axis tick labels.}

\item{labelAdj}{Numeric denoting placement of label to tick. Is passed to
second value in par("mgp").}

\item{interval}{Integer > 1. Sets interval at which axis tick labels will
be displayed.}

\item{offset}{Integer (positive) to offset where tick labels are placed.}

\item{format}{Logical. True to employ \code{AxisFormat} to reduce 0s before
or after decimal when axis tick lables are large or small numbers.
Default is False.}
}
\value{
Null (if format `FALSE`) otherwise \code{AxisFormat} output.
}
\description{
Creates custom axis tick labels for \code{Plot}.
}
\examples{
# coming soon.
}
