% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{AxisFormat}
\alias{AxisFormat}
\alias{AxisFormat.numeric}
\title{Customize Axis Tick Labels}
\usage{
AxisFormat(x, ...)

\method{AxisFormat}{numeric}(x, ...)
}
\arguments{
\item{x}{Vector (for now numeric only) typically supplied from
\code{axTicks()}.}

\item{...}{Currently not used}
}
\value{
A list with new re-formatted labels and function \code{AxisTitle} for
  customizing axis label.
}
\description{
\code{AxisFormat} (for now the numeric method) customizes axis
  tick labels by reducing tick labels with many characters to the bare
  minimum. For example, 0.015 becomes 15 or 1500 becomes 1.5. The returned
  function \code{AxisTitle} allows for customizing the axis label with `x
  10^power`, where `power` is either negative or positive depending upon
  input.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: method for object class `numeric`.
}}

\examples{
AxisFormat(c(0, 0.015, 0.020, 0.025, 0.030))
}
