% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{BackCalcLength}
\alias{BackCalcLength}
\title{Uses Dahl-Lea method to back-calculate length at age.}
\usage{
BackCalcLength(data, lCap, rCap, rAtAge)
}
\arguments{
\item{data}{Dataframe where each row is a single fish & contains
fields below.}

\item{lCap}{Field name in data containing (numeric) length at capture.}

\item{rCap}{Field name in data containing (numeric) scale radius at capture.}

\item{rAtAge}{Field (column) names or column numbers (easier) in \code{data}
containing scale radius at each age \emph{i}.}
}
\value{
A dataframe with two numeric fields: \code{len} & \code{age}. NA
   values have been removed.
}
\description{
\code{BackCalcLength} employs the Dahl-Lea algorithm to
   back-calculate length at age given specific data from the analysis
   of hard parts (e.g., scales). Assumes for now first age is 1.
}
\examples{
# code here
}
\references{
Lea, E. 1910. On the methods used in the herring-investigations.
            Publ. Circonst. Cons. perm. int. Explor. Mer 108(1):14–22.

            Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman
            & Hall/CRC, Boca Raton, FL.
}
