% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{CatchCutoffMatrix}
\alias{CatchCutoffMatrix}
\title{Create Annual (or Sampling Year) Length Cutoff Matrix.}
\usage{
CatchCutoffMatrix(relYear, minLen, maxLen, vb)
}
\arguments{
\item{relYear}{Numeric. Release (tagging) or sampling 4-digit year.}

\item{minLen}{Numeric. Minimum length, typically per regulations in place
during `relYear`.}

\item{maxLen}{Numeric. Maximum length, typically per regulations in place
during `relYear`.}

\item{vb}{Function. Output of \code{GetVBGM()}.}
}
\value{
Matrix with five columns (RelYear, MinLen, MaxLen, MinAge, MaxAge),
  where MinAge & MaxAge calculated using length inputs & `vb` function.
}
\description{
Because we cannot easily age sturgeon, length acts as a
  surrogate. Here, given min length and max length (usually from annual
  extant regulations) and VBGM, we calculate corresponding min age & max age.
  These ages then in turn are used to create length cutoff matrices.
}
\examples{
# coming soon.
}
