% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{CatchLengthCutoff}
\alias{CatchLengthCutoff}
\alias{CatchLengthCutoff.CatchCutoffMatrix}
\title{Create Length Cutoff for White Sturgeon Catch.}
\usage{
CatchLengthCutoff(x, ...)

\method{CatchLengthCutoff}{CatchCutoffMatrix}(x, ..., year)
}
\arguments{
\item{x}{Currently object of class `CatchCutoffMatrix`.}

\item{...}{Passed on to \code{link{[base]{round}}} `digits` parameter.}

\item{year}{Numeric. Year in which length data was collected (likely
considered a recapture year).}
}
\value{
List with two items. `CLC` or catch length cutoff matrix &
  `CountCatch()`, a function that accepts length data and returns counts
  given criteria (lower and upper bounds) set forth in `CLC`.
}
\description{
For White Sturgeon, the `C` in MC/R is collected over at least
  several years. To ensure consistency with length (size) of the estimate
  year, we limit `C` lengths using VBGM. (A better explanation is needed.)
}
\section{Methods (by class)}{
\itemize{
\item \code{CatchCutoffMatrix}: for object class `CatchCutoffMatrix`.
}}

\examples{
# coming soon.
}
