% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{ChapmanRobson}
\alias{ChapmanRobson}
\title{Calculates survival rate given age frequency.}
\usage{
ChapmanRobson(ages, n, N0 = 2)
}
\arguments{
\item{ages}{Numeric vector of ages.}

\item{n}{Numeric vector (length of ages) containing count (n) at each age.}

\item{N0}{Single value denoting starting age.}
}
\value{
Survival rate estimate \eqn{\hat{S}} with variance
   Var(\eqn{\hat{S}}). Also returns sum{ages used * count} and
   sum{frequency of all ages used}
}
\description{
\code{ChapmanRobson} adapted from Ricker 1975, eqn. 2.5 (2.6 for variance).
   Method is similar to using catch curve with age frequency.
}
\examples{
# ChapmanRobson(ages = , n = , N0 = 3)
}
\references{
Ricker, W.E. (1975). Computation and Interpretation of
   Biological Statistics of Fish Populations. Bulletin of the Fisheries
   Research Board of Canada, Bulletin 191, Ottawa.
}
