% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_returns.R
\name{CheckCardTag}
\alias{CheckCardTag}
\title{Check Angler-Reported Reward Disc Number.}
\usage{
CheckCardTag(x)
}
\arguments{
\item{x}{Character. The reported tag number or field containing such data.
If field, then will need to employ \code{lapply} or \code{vapply} or the
like, as function is not vectorize.}
}
\value{
Name character vector lenght = 2. `Tag` (either the tag number or
  blank) and `Desc` or a one- or two-word description of `Tag`.
}
\description{
The Sturgeon Report Card provides a field for reporting a
  disc tag should the kept or released White Sturgeon possess one. A number
  of anglers report partial tag numbers or other info (e.g., $50) making
  difficult the process of matching with CDFW tag release data. This function
  sorts through these reported "numbers" to decipher what is or might be a
  real tag to facilitate matching with tag release data.
}
\examples{
# coming soon
}
