% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_length.R
\name{ConvertLength}
\alias{ConvertLength}
\title{Converts Total Length to Fork Length or Fork Length to Total Length.}
\usage{
ConvertLength(data, tl, fl, sp, convertTo = c("TL", "FL"))
}
\arguments{
\item{data}{Dataframe containing (at least) length (fork & total) & species.}

\item{tl}{Field name (not quoted) in \code{data} that holds total length.}

\item{fl}{Field name (not quoted) in \code{data} that holds fork length.}

\item{sp}{Field name (not quoted) in \code{data} that holds species.}

\item{convertTo}{Character scalar option of either "TL" or "FL", the length
type converting to.}
}
\value{
A numeric vector with length = \code{nrow(data)} where data are total
  length or fork length per choice in \code{convertTo}. Where converting to
  length already exists, that length it used rather than converting. If both
  TL & FL are NA, then returned length is also NA.
}
\description{
Mostly this function applies to California-based sturgeon and
  mostly because in 2013 regulations changed from total length to fork length
  (cm). The CDFW now measures fork length, so this function facilitates
  conversion between years allowing for more consistent analytics.
}
\note{
Conversions use linear regression models (one for White Sturgeon
  another for Green Sturgeon). See References for source of models.

  \deqn{FL_{WST}=-1.2162 + (0.9036 * TL_{WST})}
  \deqn{TL_{GST}=-4.6131 + (1.1374 * FL_{GST})}
}
\examples{
# coming soon
}
\references{
coming soon for equations
}
