% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{CountNets}
\alias{CountNets}
\title{Counts Fyke Traps or Gill Nets Fished During Desired Period.}
\usage{
CountNets(netNum)
}
\arguments{
\item{netNum}{Numeric or character value given net set or (fyke) trap number.}
}
\value{
A list with the following values:
\itemize{
  \item RecordCount (number of values)
  \item NetCount (record count for unique net numbers)
  \item NetCountDup (count of traps or nets fished more than once in a day)
  \item UniqueNets (trap or net numbers - unique)
  \item DupNets (trap or net numbers - non-unique)
}
}
\description{
A simple function to count fyke traps or gill nets fished during
  specified period. Period is specified (created) by user, not within this
  function.
}
\examples{
# coming soon
}
