% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap-methods.R
\name{CountRecap}
\alias{CountRecap}
\alias{CountRecap.numeric}
\alias{CountRecap.xtabs}
\title{Count Tag Recaptures For Each Release Year.}
\usage{
CountRecap(x, ...)

\method{CountRecap}{numeric}(x, ..., n, recYears, nRecaps)

\method{CountRecap}{xtabs}(x, ..., n)
}
\arguments{
\item{x}{Object of either class `numeric` or class `xtabs`.}

\item{...}{Passed to other methods. Currently not used.}

\item{n}{Integer giving the number in years for how long to count recaptures.}

\item{recYears}{Numeric vector providing 4-digit recapture years.}

\item{nRecaps}{Numeric vector providing number of recaptures per recapture
year.}
}
\value{
Numeric vector with the following items.
   \describe{
      \item{NRecYears}{count of recapture years inlcuded in summary}
      \item{CutoffYear}{year in which counting recaptures stopped}
      \item{NRecaps}{number of recaptured tags for release year}
   }
}
\description{
Count of tag recaptures for each release year is required to
   estimate absolute abundance (i.e., the `R` in MC/R). \code{CountRecap}
   methods provide a convenient way to get such counts along with the
   cutoff recapture year and count (n) of how many recapture years were
   included.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: for class numeric.

\item \code{xtabs}: for class xtabs.
}}

\examples{
# coming soon.
}
