% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creel.R
\name{CreelType}
\alias{CreelType}
\title{Assign River Creel or Bay Creel Type Based on Census Date.}
\usage{
CreelType(date)
}
\arguments{
\item{date}{Class Date or POSIX; sampling date (YYYY-MM-DD)}
}
\value{
List with Year, Month, and Type items, where Year is +1 if
   Month = 12 and Type = 'rvr'.
}
\description{
Striped Bass creel census is bifurcated. River creel from
   Dec 16 - Jun 15 and Bay Creel from Jun 16 - Dec 15. This function
   assigns one or the other based on `date` argument.
}
\examples{
# coming soon.
}
