% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{CumsumNA}
\alias{CumsumNA}
\title{Cumulative Sum For Numeric Vectors With NA.}
\usage{
CumsumNA(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Cumulative sum of `x`, where any NA values remain NA.
}
\description{
\code{\link[base]{cumsum}} returns NA for all values at and
   after first NA. \code{CumsumNA} 'ignores' NA values.
}
\examples{
CumsumNA(c(1:4, NA, 5:10))
}
