% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{DescStat}
\alias{DescStat}
\title{Calculates descriptive statistics of \code{x}.}
\usage{
DescStat(x)
}
\arguments{
\item{x}{A numeric vector, preferably of length > 1.}
}
\value{
A list containing n (with & without NA values; NAll & N),
   min, max, mean, median, and variance.
}
\description{
\code{DescStat} provides basic descriptive statistics of
   numeric input. Will remove all NA values prior to calculations &
   will issue warning when doing so.
}
\note{
To get standard error run \code{sqrt(Var / N)}.
}
\examples{
DescStat(1:10)
DescStat(c(runif(10), NA))
}
