% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{EffortStu}
\alias{EffortStu}
\title{Calculates Fishing Effort for Sturgeon Trammel Nets.}
\usage{
EffortStu(data, sStart, sEnd, rStart, rEnd, netLength)
}
\arguments{
\item{data}{Dataframe containing (at least) net set & net retrieval time and
net length (typically in fathoms).}

\item{sStart}{date-time object. Time first bit of webbing entered water.}

\item{sEnd}{date-time object. Time last bit of webbing entered water.}

\item{rStart}{date-time object. Time first bit of webbing pulled from water.}

\item{rEnd}{date-time object. Time last bit of webbing pulled from water.}

\item{netLength}{Numeric value given length of net fishing
(e.g., 100 fathoms).}
}
\value{
list with the following elements:
   \itemize{
      \item Effort (fishing effort)
      \item TimeFished (in hours; total fishing hours)
      \item SoakTime (in hours)
      \item TimeUnit (characater)
      \item NetLength (the variable in \code{data} containing net length)
   }
}
\description{
\code{EffortStu} will calculate fishing effort for each net set
   given fishing times and (trammel) net length. Fishing effort is the
   denominator when calculating CPUE (or catch per unit effort).
}
\examples{
# coming soon.
}
