% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitation.R
\name{Exploitation}
\alias{Exploitation}
\title{Calculcates adjusted exploitation rate \eqn{\mu\prime}.}
\usage{
Exploitation(rr, rt, repRate, tagL, tagM)
}
\arguments{
\item{rr}{Number of non-reward tags reported (positive integer).}

\item{rt}{Number of non-reward tags released (positive integer).}

\item{repRate}{Reporting rate \code{RepRate} from \code{ReportingRate}.}

\item{tagL}{Fraction of tags lost (default ??).}

\item{tagM}{Mortality (as a fraction) from tagging (default ??).}
}
\value{
List containing number returned, number released (tagged),
   reporting rate, raw exploitation rate mu, and adjusted exploitation
   rate \eqn{\mu\prime}.
}
\description{
Adjusted exploitation compensates for lack of angler-
   reporting (i.e., via reporting rate), tag loss, and mortality
   due to tagging. If tag loss or tagging mortality are unknown,
   values from literature can be used.

   \deqn{\mu\prime=\frac{\mu}{(repRate \times (1-tagL) \times (1-tagM))}}
}
\examples{
# numbers for example only
Exploitation(rr = 5, rt = 175, repRate = 0.5, tagL = 0.05, tagM = 0.05)
}
\references{
Kevin A. Meyer & Daniel J. Schill (2014) Use of a Statewide
   Angler Tag Reporting System to Estimate Rates of Exploitation
   and Total Mortality for Idaho Sport Fisheries, North American
   Journal of Fisheries Management, 34:6, 1145-1158,
   DOI: 10.1080/02755947.2014.951803
}
