% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baystudy.R
\name{FlagTows}
\alias{FlagTows}
\title{A "Where Clause" or Filter for Bay Study Tows.}
\usage{
FlagTows(data, station, year, survey, tow)
}
\arguments{
\item{data}{Dataframe of Bay Study data with at least station, year, survey,
and tow.}

\item{station}{Name of field in data holding station info.}

\item{year}{Name of field in data holding year (i.e., extracted from date).}

\item{survey}{Name of field in data holding survey number.}

\item{tow}{Name of field in data holding tow nubmer.}
}
\value{
List containing flagged records (-1 = flagged; O = OK to use) and
  number of records flagged.
}
\description{
Flags "valid" tows per select criteria.  Typically, excluded are
  tows 52, 53, and 58; station #141 (only sampled a few months, then
  dropped); and stations #317 and #319 in 1980 and 1981-01 (split surveys,
  used stations #324 and #326, which were the same locations, sampled 1-2
  weeks later).
}
\examples{
# coming soon.
}
\references{
Bay Study metadata (insert link).
}
