% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{Frequency}
\alias{Frequency}
\alias{summary.Frequency}
\alias{print.Frequency}
\title{Create a frequency distribution of \code{x} by \code{binWidth}.}
\usage{
Frequency(x, binWidth, xRange = NULL)

\method{summary}{Frequency}(object, ...)

\method{print}{Frequency}(x, ...)
}
\arguments{
\item{x}{A postive numeric vector to be binned according to \code{binWidth}.}

\item{binWidth}{A numeric scalar (integer) defining the bin size.}

\item{xRange}{A numeric vector of length 2, supplying 'from' and 'to' for
customized x-axis range. Default NULL. Recommended for consistent
comparison of multiple histograms.}

\item{object}{Object of class \code{Frequency}.}

\item{...}{Passed on to other methods.}
}
\value{
A list (of class \code{Frequency} & class
  \code{\link[graphics]{hist}}) containing proportion & count frequencies.
}
\description{
\code{Frequency} is a custom wrapper (mostly) for
  \code{\link[graphics]{hist}}. It simplifies input to only data and bin
  width. Bins are always left included but not right, as in \code{`[a, b)`}.
  \code{Frequency} removes \code{NA} values (with warning) prior to binning.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: \code{summary} method (See notes).

\item \code{print}: \code{print} method (See notes).
}}

\note{
Method \code{summary} written for class Frequency. Returns dataframe
  with fields Bin, Count, Density, Proportion. Print method also writted for
  class Frequency.
}
\section{Values}{


  \describe{ \item{others}{see \code{\link{hist}}}. \item{xna}{number of NA
  values removed}. \item{xname}{deparsed name of \code{x}, replaces
  \code{hist} \code{xname}} \item{binw}{value passed to \code{binWidth}}
  \item{xstats}{ function when run will yield descriptive statistics of
  \code{x} } }
}

\examples{
#Frequency(trammel_catch[["FL"]], binWidth = 5)
}
