% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaptures.R
\name{GetStuRecap}
\alias{GetStuRecap}
\title{Retrieve Dataframe of Sturgeon Tag Recaptures.}
\usage{
GetStuRecap(data, ...)
}
\arguments{
\item{data}{Dataframe containing CDFW mark-recapture Sturgeon data.}

\item{...}{Field names (not quoted) in `data` where each is named (see
details).}
}
\value{
List with four items: (1) `TagNum` - the tag number of the
  recaptured tag; (2) `TagType` - describes recapture (i.e., a
  recapture; a retag; etc.); (3) `RecordCount` - count of records
  by tag type & total count; (4) `Data` - a list with appropriate
  release & recapture info for each recaptured tag. List is class
  `GetStuRecap`.
}
\description{
Get description here
}
\details{
Items in `...` must be named accordingly.
\describe{
  \item{type}{sturgeon type, i.e. Tagged, Recapture, or Not Tagged;
              usually found in `StuType` field in `Sturgeon` data}
  \item{shed}{field with info on whether the sturgeon shed or
              possibly shed the disc tag}
  \item{retag}{field with info on whether the sturgeon was retagged}
  \item{dn}{field with original disc tag number}
  \item{pn}{field with original PIT number}
  \item{od}{field with old disc tag number}
  \item{op}{field with old PIT number}
}
}
\note{
Admittedly, `StuType` is a poorly named field & somewhat
  misleading (e.g., can be confused with species). This field
  essentially describes the table in which data live in the database
  (tagged, not tagged, or recaptured). Some tagged fish can be
  classified as recaptured if the fish had a previous tag and was (thus)
  retagged. \code{GetStuRecap} and other internal functions capture
  all recaptured fish (including those retagged or those fish with only
  PIT tags, no disc tags).
}
\examples{
# coming soon
}
