% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{GetVBGM}
\alias{GetVBGM}
\title{Create Function to Apply von Bertalanffy Growth Model (VBGM).}
\usage{
GetVBGM(Linf, K, t0)
}
\arguments{
\item{Linf}{Numeric. Asymptotic length parameter from VBGM.}

\item{K}{Numeric. Estimated growth rate parameter from VBGM.}

\item{t0}{Numeric. Age at zero size paramater from VBGM.}
}
\value{
Returns a function to supply length (default) or age (`x`) which then
  returns age (default) or length.
}
\description{
Function returns a function which is then used to get length at
  age or age at length given VBGM parameters. The VBGM parameters are
  returned with the second function for use in other applications.
}
\note{
Paramater `output`: Character. Paramter from returned function accepts
  either `laa` or `aal` (length-at-age or age-at-length; default `laa`).
}
\examples{
# coming soon
}
