% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{GroupYear}
\alias{GroupYear}
\alias{GroupYear.Date}
\alias{GroupYear.data.frame}
\title{Extract 4-digit Year Based on Starting Month.}
\usage{
GroupYear(x, ...)

\method{GroupYear}{Date}(x, ..., startMon)

\method{GroupYear}{data.frame}(x, year, mon, ..., startMon, direction = c("forward", "backward"))
}
\arguments{
\item{x}{Object of class `Date` or `data.frame`.}

\item{...}{Currently not used.}

\item{startMon}{Integer scalar value between 1 and 12.}

\item{year}{Field name in `x` when `x` is `data.frame` containing numeric
4-digit year.}

\item{mon}{Field name in `x` when `x` is `data.frame` containing numeric 1-
to 2-digit month (i.e., 1-12).}

\item{direction}{Character of either "forward" or "backward" denoting whether
year is moved forward (default; `+`) or backward (`-`).}
}
\value{
Numeric 4-digit year based on starting month.
}
\description{
\code{GroupYear} groups a year according to a starting month.
  This function is useful for grouping values in year x & year x + 1 crossing
  over calendar year (e.g., months Oct-Sep).
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: Method for object class `Date`.

\item \code{data.frame}: Method for object class `data.frame`.
}}

\note{
\code{GroupYear} not yet equipped to control `end` month, so result is
  full 12 months based on starting month.
}
\examples{
# coming soon
}
