% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{HoursSb}
\alias{HoursSb}
\title{Calculates Fishing Time in Hours.}
\usage{
HoursSb(sTime, eTime)
}
\arguments{
\item{sTime}{A date-time object denoting fishing start time.}

\item{eTime}{A date-time object denoting fishing end time.}
}
\value{
Value(s) of class numeric given eTime - sTime.
}
\description{
A simple function using \code{base::difftime} to calculate in
  hours time fished for fyke traps or gill nets. For convenience, output is
  converted to numeric rather than class "difftime."
}
\examples{
# coming soon
}
