% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_length.R
\name{LengthGroup}
\alias{LengthGroup}
\title{Assign Each Fish to Length Group.}
\usage{
LengthGroup(
  data,
  year,
  species = NULL,
  tl = NULL,
  fl = NULL,
  regTable = c("su", "sb")
)
}
\arguments{
\item{data}{Dataframe containing (at least) fish length & catch date (year).}

\item{year}{Field name (not quoted) in \code{data} that holds capture
(collection) year.}

\item{species}{Field name (not quoted) in \code{data} that holds species.}

\item{tl}{Field name (not quoted) in \code{data} that holds total length.}

\item{fl}{Field name (not quoted) in \code{data} that holds fork length.}

\item{regTable}{Dataframe of regulation changes with at least the following
fields: "MaxYear"; "Species"; "MinLen"; "MaxLen"; and "LenType". Default
is `su` and gets internal `RegsSturgeon`. Other choice is `sb`
and will get internal `RegsStriper`.}
}
\value{
A character vector with length = \code{nrow(data)} & possible values
  of 'sub', 'leg', 'ovr', 'unk', or 'not'.
}
\description{
Assigns for each fish 'sub' (sub-legal), 'leg' (legal), 'ovr'
  (over-legal) based on fishing regulations and measured fork length or total
  length (cm). If length is NA, then 'unk' (unknown) is used instead. Special
  case for Green Sturgeon using 'not' post 2006, when it be came illegal to
  fish for (and keep) said species.
}
\examples{
# coming soon
}
