% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tables.R
\docType{data}
\name{lkp_striper}
\alias{lkp_striper}
\title{Look-up Table for Striped Bass Mark-Recapture Data.}
\format{An object of class \code{list} of length 5.}
\usage{
lkp_striper
}
\description{
A list containing look-up tables for various fields found in
  Striped Bass data. These look-up tables assist in data clean up & make
  analytics easier to summarize. For example, a fish conditions of "Good" or
  "good" in the database will cause issues with a case-senstive language like
  R. These look-up tables will rectify this issue.
}
\keyword{datasets}
