% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{PanelCorr}
\alias{PanelCorr}
\title{Add Correlation and Smoother to `graphics::pairs()` Plot.}
\usage{
PanelCorr(out = c("points", "text"))
}
\arguments{
\item{out}{Character. Either "points" (default) or "text allowing for either
points with smoother or text (as *r* & p-value). Pass this function to
either `lower.panel` or `upper.panel` in `graphics::pairs()`.}
}
\value{
Function suited for use as described above.
}
\description{
Allows one to easiliy display smoother curve & correlation (with
  p-value) on `graphics::pairs()` plot.
}
\examples{
# coming soon
}
