% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{PanelLM}
\alias{PanelLM}
\title{Add Linear Regression (Line & Ouput) to `graphics::pairs()` Plot.}
\usage{
PanelLM(out = c("points", "text"))
}
\arguments{
\item{out}{Character. Either "points" (default) or "text allowing for either
points with linear regression line or text (e.g., slope; intercept). Pass
this function to either `lower.panel` or `upper.panel` in
`graphics::pairs()`.}
}
\value{
Function suited for use as described above.
}
\description{
Allows one to easiliy display linear regression & some output
  (e.g., slope; intercept; p-value) on `graphics::pairs()` plot.
}
\examples{
# coming soon
}
