% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{PetersenEst}
\alias{PetersenEst}
\alias{PetersenEst.numeric}
\alias{print.PetersenEst}
\title{Calculate Petersen Abundance Estimates.}
\usage{
PetersenEst(x, ...)

\method{PetersenEst}{numeric}(x, CC, RR, ..., alpha = 0.05)

\method{print}{PetersenEst}(x, ...)
}
\arguments{
\item{x}{Numeric. Number of fish marked (tagged).}

\item{...}{Currently not used.}

\item{CC}{Numeric. Number of fish caught or sampled.}

\item{RR}{Numeric. Number of recaptured fish (tags) in the sample.}

\item{alpha}{Numeric scalar used to calculated \code{z-value} for defining
confidence intervals (e.g., 95\%). Default at 0.05 (or 95\%).}
}
\value{
A list (of class \code{PetersenEst}) containing items in
  \code{Values} section.
}
\description{
\code{PetersenEst} calculates abundance estimates using
  mark-recapture data. Function is designed following equations detailed in
  Ricker 1975 (see References). Currently, we use Bailey's modified
  equation (see eqn 3.9, page 79 in Ricker 1975).
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: for object class `numeric`.

\item \code{PetersenEst}: \code{print} method (See notes).
}}

\note{
Method \code{print} written for class PetersenEst returns estimate
  with confidence interval.
}
\section{Values}{


  \describe{
    \item{N}{absolute abundance estimate}.
    \item{Var}{variance of N}.
    \item{CI}{confidence interval as SE * Z}.
    \item{SE}{standard error as \code{sqrt(Var)}}.
    \item{Level}{CI level as a result of Z; e.g., 95\%}.
    \item{Z}{given as \code{qnorm(1 - alpha / 2)},
             where alpha = 0.05 for 95\% CI}.
   }
}

\examples{
# coming soon
}
\references{
Ricker, W.E. (1975). Computation and interpretation of biological
  statistics of fish populations. Bulletin of the Fisheries Research Board
  of Canada, Bulletin 191, Ottawa. http://www.dfo-mpo.gc.ca/Library/1485.pdf
}
