% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{plot.Frequency}
\alias{plot.Frequency}
\title{Generate base R histogram from object of class \code{\link{Frequency}}.}
\usage{
\method{plot}{Frequency}(
  x,
  ...,
  med = TRUE,
  xTL = TRUE,
  yTL = TRUE,
  xTitle = TRUE,
  yTitle = TRUE,
  addN = TRUE,
  maxY = NULL
)
}
\arguments{
\item{x}{An object of class Frequency (& histogram).}

\item{...}{Passed on to other methods.}

\item{med}{Logical. If true (default) plot will display vertical line
denoting median value.}

\item{xTL}{Logical. If true (default) plot will include x-axis tick labels.}

\item{yTL}{Logical. If true (default) plot will include y-axis tick labels.}

\item{xTitle}{Logical. If true (default) plot will include x-axis title.}

\item{yTitle}{Logical. If true (default) plot will include y-axis title.}

\item{addN}{Logical. If true (default) plot will include `N` value outside
plot area (top left).}

\item{maxY}{Numeric. If null (default) plot uses \code{max(x[["density"]])}.}
}
\value{
Displays a histogram using frequency item in \code{x}.
}
\description{
Plot a histogram using \code{hist} (esentially calling next
  method on \code{plot}, which calls \code{plot.histogram}). Display contains
  sample size along with some descriptive statistics.
}
\examples{
#freq <- Frequency(trammel_catch[["FL"]], binWidth = 5)
#plot(freq)
}
