% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{Plot}
\alias{Plot}
\title{Create Custom Empty Base R Plot.}
\usage{
Plot(
  data = NULL,
  x,
  y,
  ...,
  yerr = NULL,
  subset = NULL,
  x0 = FALSE,
  y0 = FALSE,
  adjUsr = 0.9,
  xAsFactor = FALSE
)
}
\arguments{
\item{data}{A dataframe with at least `x` & `y` data fields.}

\item{x}{A vector if `data` is NULL or field name in `data` for the x axis.}

\item{y}{A vector if `data` is NULL or field name in `data` for the y axis.}

\item{...}{Passed to other methods. Currently not used.}

\item{yerr}{A vector if `data` is NULL or field name in `data` for the
y axis error bars. Currently not such parameter for x error bars.}

\item{subset}{Logical expression to be evaluated in `data` or for `x` & `y`.}

\item{x0}{Logical. True if x-axis minimum value should be 0. Default False.}

\item{y0}{Logical. True if y-axis minimum value should be 0. Default False.}

\item{adjUsr}{Numeric. User to ajust \code{par("usr")} mostly to smaller
size, as the default is 0.90 (thus a 10\% reduction).}

\item{xAsFactor}{Logical. True if x-axis should be a factor. Defaul = FALSE.}
}
\value{
A list (to be detailed later).
}
\description{
\code{Plot} creates an empty plot with axes of suitable limits
   so as to include all data points. Plot created is then ready for points
   or lines or bars (etc.) overlay.
}
\examples{
# coming soon.
}
