% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{points.plot_empty}
\alias{points.plot_empty}
\title{Add Data to \code{Plot}.}
\usage{
\method{points}{plot_empty}(
  x,
  type = "p",
  ...,
  grid = TRUE,
  xRng = FALSE,
  yRng = FALSE,
  nx = 1,
  ny = 1
)
}
\arguments{
\item{x}{An object of class `plot_empty`.}

\item{type}{Scalar character of either `p` (points), `l` (lines), or `h`
(bars); see `help("plot.default")` for more.}

\item{...}{Other parameters passed to \code{points}.}

\item{grid}{Logical. True (default) will display grid on `Plot`.}

\item{xRng}{Logical. False (default) keeps default `xaxp`. True uses
c(range(x), diff(range(x))).}

\item{yRng}{Logical. False (default) keeps default `yaxp`. True uses
c(range(y), diff(range(y))).}

\item{nx}{Integer (default 1) affecting number of vertical lines in grid.}

\item{ny}{Integer (default 1) affecting number of horizontal lines in grid.}
}
\value{
Nothing at the momement.
}
\description{
Add data points (lines, etc.) to empty plot with or without
   background plus grid lines.
}
\examples{
# coming soon.
}
