% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_year.R
\name{print.return_year}
\alias{print.return_year}
\title{Print Method for \code{TagRetYear}.}
\usage{
\method{print}{return_year}(x, ...)
}
\arguments{
\item{x}{Object of class `return_year`.}

\item{...}{Currently not used.}
}
\value{
Count of NA values and count of negative values, if any.
}
\description{
Print Method for \code{TagRetYear}.
}
\examples{
# coming soon
}
