% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{ReadCaptionFile}
\alias{ReadCaptionFile}
\title{Read Caption Text File.}
\usage{
ReadCaptionFile(file)
}
\arguments{
\item{file}{Character. Name of file containing captions
(along with relevant file path).}
}
\value{
Named character string, where names are caption
   heading (e.g., 'table1').
}
\description{
Makes easier the task of writing captions (and then reading)
   captions from a separate file, rather than within `.Rmd`. The text file
   can contain all captions (e.g., figures; tables), so long as each caption
   is prefixed with '<caption_name><caption_number>: ' (e.g., "Table 1: ").
}
\examples{
# comin soon.
}
