% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadRDSFiles}
\alias{ReadRDSFiles}
\title{Read All `.rds` Files from a Single Directory.}
\usage{
ReadRDSFiles(fileDir, envir = .GlobalEnv)
}
\arguments{
\item{fileDir}{Character. Name of directory (with path) containing `.rds`
files.}

\item{envir}{Character. Environment in which to save loaded .`rds` files.
Default is global environment.}
}
\value{
Names of `.rds` files loaded as dataframes to `envir`.
}
\description{
\code{ReadRDSFiles} is basically a wrapper for
  \code{base::readRDS} and makes convenient the loading at onces of multiple
  files.
}
\examples{
# none at this time.
}
