% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_ratio.R
\name{SexRatio}
\alias{SexRatio}
\alias{SexRatio.sex_striper}
\title{Calculate Count and Proportion of Male to Female}
\usage{
SexRatio(x, ...)

\method{SexRatio}{sex_striper}(x, ..., digits = NULL)
}
\arguments{
\item{x}{Object of class `sex_striper`.}

\item{...}{Currently not used.}

\item{digits}{Integer for rounding proportion.}
}
\value{
List with two elements (frequency & proportion).
}
\description{
Simplifies calculation of Striped Bass sex ratio. For now, only
  available for object class `sex_striper`.
}
\section{Methods (by class)}{
\itemize{
\item \code{sex_striper}: for object class `sex_striper`.
}}

\examples{
# coming soon.
}
