% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Split}
\alias{Split}
\title{Create List Column for Analytics.}
\usage{
Split(data, subset = NULL, vars = NULL, splitVars, drop = FALSE)
}
\arguments{
\item{data}{A dataframe to be subsetted (optional) & split accordingly.}

\item{subset}{A Boolean expression evaluated in `data`.}

\item{vars}{Desired field names (not quoted) in `data`.}

\item{splitVars}{Field names (or names) in `data` on which to split.}

\item{drop}{Logical. Default FALSE (should non-ocurring levels be dropped).}
}
\value{
A data.frame with `Data` as a list column along with fields in
  `splitVars`. Row count based on `splitVars`.
}
\description{
\code{Split} creates a list-column dataframe for ease of
  analytics. The `Data` column contains the results of
  \code{\link[base]{split}}.
}
\examples{
# coming soon
}
