% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap-methods.R
\name{StuRecap}
\alias{StuRecap}
\alias{StuRecap.RecapData}
\alias{StuRecap.StuRecap}
\title{Get Sturgeon Recapture Summary Data.}
\usage{
StuRecap(x, ...)

\method{StuRecap}{RecapData}(x, ...)

\method{StuRecap}{StuRecap}(x, ..., species = "w", lencat = NULL, minDAL = -1L)
}
\arguments{
\item{x}{Object of class `RecapData` or class `StuRecap`.}

\item{...}{Passed to other methods or functions. Currently not used.}

\item{species}{Scalar characater either 'w' (default) or 'g'. Defines
species either White Sturgeon or Green Sturgeon.}

\item{lencat}{Scalar character with choices 's', 'l', or 'o'. Defines
length category for desired subsetting. Sub-legal (s), legal (l), or
over-legal (o) are based on length (and existing regulations) at time
fish was tagged. Default (`NULL`) includes all length categories.}

\item{minDAL}{Integer. Defines minimum days-at-large post tag & release.
Oftentimes it is best (for minimizing biases) when tagged fish mixes
with non-tagged fish. Limiting to at least ~30 days-at-large to define
a recapture may mitigate biases. Default value (-1L) does not subset
for DAL.}
}
\value{
List with nine items.
 \describe{
   \item{RelYr}{4-digit release year}
   \item{RecYr}{4-digit recapture year(s)}
   \item{RelDate}{release date (YYYY-MM-DD)}
   \item{RecDate}{recapture date(s) (YYYY-MM-DD)}
   \item{DAL}{days at large; RecDate - RelDate}
   \item{LenCat}{length category at tagging: sub; leg; ovr}
   \item{Species}{species (either white or green)}
   \item{DupTag}{see note}
   \item{NRecs}{number of records in recapture data}
 }
}
\description{
The CDFW Sturgeon Mark-Recapture Study annually affixes a single
  disc tag (mark) to many sturgeon. Sometimes marks from within season or
  previous seasons are recaptured and the disc number recorded (along with
  other pertinent data). \code{StuRecap()} makes easier the process of
  getting (for each tag) release year, recapture year, & days-at-large (DAL),
  along with other metrics. Summary data are then used to construct a
  frequency table of sorts with the count of tag recaptures (by recapture
  year) for each release year.
}
\section{Methods (by class)}{
\itemize{
\item \code{RecapData}: for object class `RecapData`.

\item \code{StuRecap}: for object class `StuRecap`.
}}

\note{
In 2015, we inadvertently released duplicate tag numbers released in
   2011 & 2012. To date, duplicate tags `HH2141` & `HH2145` are the only two
   that have been recaptured. These are handled specially within
   \code{StuRecap.RecapData()} to account for two separate release dates.
}
\examples{
# coming soon.
}
