% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_count.R
\name{TagCount}
\alias{TagCount}
\title{Get Tag Count.}
\usage{
TagCount(data, .by, .subset, total = FALSE)
}
\arguments{
\item{data}{Dataframe with tag (return or release) data.}

\item{.by}{Single field name (non-quoted) in `data` on which to aggregate
record (tag) count.}

\item{.subset}{Logical. Booleans on which to count. Any grouping of Booleans
evaluating to `TRUE` count as 1.}

\item{total}{Logical. Option to include `total` count. Default `FALSE`.}
}
\value{
Dataframe with fields `Val` (as in tag value) & `N` (count).
}
\description{
\code{TagCount} in reality is not specific to tags, rather this
  function will yield record count by some variable `v` for which `.subset`
  yields true. Ideally we use \code{TagCount} to count tags (released or
  returned) by tag value (i.e, non-reward and [or] dollar values).
}
\examples{
# coming soon
}
