% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_year.R
\name{TagRetYear}
\alias{TagRetYear}
\title{Calculate Return Year for Angler Tag Returns.}
\usage{
TagRetYear(data, dcap, drel)
}
\arguments{
\item{data}{Dataframe containg angler tag return information (at the very
least tag, date captured, and date released).}

\item{dcap}{Field name (unquoted) in `data` containing capture date.}

\item{drel}{Field name (unquoted) in `data` containing release date.}
}
\value{
List. See `values`.
}
\description{
Angler tag returns are classified based on the difference (in
  days) between the reported capture date and recorded release (tagging)
  date. This difference is converted to years (by dividing by 365.25 [0.25
  accounts for leap year]) and is assigned as the `return year` for the
  returned tag. A tag returned within 365.25 days post released is classified
  as return year = 1, and so on. Return year is used (for example) in
  exploitation estimates, where only tags with return year = 1 are used.
}
\note{
Date fields should be of class `Date` or `POSIX`.
}
\section{values}{

   \itemize{
      \item DAL Numeric vector. Difference in days between dates.
            (`DAL` = days at large)
      \item RetYear Numeric vector. Calculated return year.
      \item CountNAsDAL Numeric scalar. Count of `NA` values.
            (i.e., where RetYear or DAL could not be calculated.)
      \item CountNegativeDAL Numeric scalar. Count where capture date
            before release date (illogical).
      \item RecordNumsNegDAL Numeric vector. If `CountNegativeDAL` not 0,
            then herein are record numbers where capture date
            before release date.
   }
}

\examples{
# coming soon
}
