% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{TimeDiffStu}
\alias{TimeDiffStu}
\title{Calculates Fishing Times of Sturgeon Trammel Net.}
\usage{
TimeDiffStu(sStart, sEnd, rStart, rEnd)
}
\arguments{
\item{sStart}{date-time object. Time first bit of webbing entered water.}

\item{sEnd}{date-time object. Time last bit of webbing entered water.}

\item{rStart}{date-time object. Time first bit of webbing pulled from water.}

\item{rEnd}{date-time object. Time last bit of webbing pulled from water.}
}
\value{
list. See 'Details' for desription of list elements.
}
\description{
The CDFW Sturgeon mark-recapture study uses trammel nets to
  catch sturgeon. Times are recorded four times: twice during deployment and
  twice during retrieval. \code{TimeDiffStu} calculates time differences
  between retrieval and deployment to ultimately calculate fishing effort.
  See 'Details' for further information.
}
\details{
Calculated differences are in hours. The returned list contains 5
  such calculations as explained below.

  \itemize{
     \item TFT (total fishing time)
     \item ST (soak time)
     \item TST (true soak time)
     \item ETST (estimated true soak time)
     \item TOEST (true or estimated soak time)
  }

  \deqn{TFT = \code{rEnd} - \code{sStart}}
  \deqn{ST = \code{rEnd} - \code{sEnd}}
  \deqn{TST = (\code{rStart} - \code{sEnd}) +
        0.5 * (\code{rEnd} - \code{rStart}) +
        0.5 * (\code{sStart} - \code{sEnd})}
  \deqn{ETST = 0.138 + (0.687 * \code{ST})}

  If \code{ST} is NA, then \code{ETST} defaults to 1.067. If \code{TST} is
  NA, then \code{ETST} is used.

  Return list also includes unit of time (currently hours) and function to
  calculcate effort using \code{TOEST}.

  Time difference calculations created by M. Donnellan. (<some reference
  to metadata here for further understanding of calcs.>)
}
\examples{
# coming soon.
}
