% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baystudy.R
\name{TowMetric}
\alias{TowMetric}
\title{Calculate Bay Study Tow Volume or Tow Area.}
\usage{
TowMetric(data, net, meter, distance)
}
\arguments{
\item{data}{Dataframe of Bay Study data with at least net type, meter value,
and distance towed.}

\item{net}{Name of field in data holding net type info.}

\item{meter}{Name of field in data holding meter info.}

\item{distance}{Name of field in data holding distance.}
}
\value{
Numeric vector of tow area or tow volume (depending upon net type).
}
\description{
\code{TowMetric} calculates tow volume if mid-water trawl or
  other and tow area if otter trawl. Does so by multiplying constants by
  (flow) meter value or tow distance.
}
\examples{
# coming soon.
}
\references{
Bay Study metadata (insert link).
}
